// dashboard-v6/src/components/api/Corpus.ts
export interface IChapterToc {
  book: number;
  paragraph: number;
  level: number;
  text: string | null;
  chapter_len: number;
  chapter_strlen: number;
  parent: number;
}

export interface IChapterTocListResponse {
  ok: boolean;
  message: string;
  data: { rows: IChapterToc[]; count: number };
}